backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.3.4
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.2.0
geopy==2.4.1
requests==2.31.0
python-dateutil==2.8.2
